+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v3.5.0
 
 - Add Cat Editor seed input [#545](https://github.com/hushenghao/AndroidEasterEggs/issues/545)
+- Add Cat Editor svg export
 - Add to favorites in Cat Editor
 - Add Bengali translation, Thanks for [@ChildishRifty7](https://github.com/childishrifty7) 's contribution
 - Rocket Launcher add icon collection selection [#554](https://github.com/hushenghao/AndroidEasterEggs/issues/554)